feat(response): create custom response struct
Closes #2

Instead of just returning a JSON string we now return a struct
with all the fields in the SocketLabs response body. I decided to
not return an error everytime we get a SocketLabs error, because
there can be errors with the request, one of the messages or one
of the addresses in one of the messages and I don't think users
will want to get actual application errors everytime one of these
errors is returned from SocketLabs.

Questions:
- We've got no headers in the response struct right now.
Should we?
- We are using the failure_derive to display the error description,
but that seems kind of odd since we are not actually implementing
the Error with it. Probably we should find a better ans more
semantic solution for that some other time.